refactor config, CLI and DAG
