+++ test/tradingo/test_config.py
@@ -3,7 +3,7 @@ import pathlib
 
 import pytest
 
-from tradingo import cli
+from tradingo import dag
 from tradingo.config import TradingoConfig
 
 
@@ -17,9 +17,9 @@ def test_config():
     )
     env.to_env()
 
-    out = cli.read_config_template(
+    out = dag.read_config_template(
         pathlib.Path("/home/rory/dev/tradingo-plat/config/tradingo/yfinance.yaml"),
         variables=os.environ,
     )
 
-    cli.DAG.from_config(out)
+    dag.DAG.from_config(out)
